requirements.txt
Flask
Flask-Login
mysql-connector-python
python-dotenv
psutil
Werkzeug
